% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access_closeConnection}
\alias{access_closeConnection}
\title{Cerrar la conexión de Access}
\usage{
access_closeConnection()
}
\value{
Se cierra la conexión.
}
\description{
Cerrar la conexión de Access
}
\examples{
con <- access_import(db)
access_closeConnection()

}
