% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access_searchTables}
\alias{access_searchTables}
\title{Tablas disponibles en Access}
\usage{
access_searchTables(con)
}
\arguments{
\item{con}{La conexión previamente establecida con \code{access_import}}
}
\value{
Un vector con el nombre de las tablas
}
\description{
Ver el nombre de las tablas disponibles en la base de datos Access
}
\examples{
library(RODBC)
# ruta de la base de datos
db <- "C:/Users/34673/Downloads/Liquidaciones2021.mdb"
# generamos la conexión
con <- access_import(db)

# vemos qué tablas hay:
access_searchTables(con)

}
