% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access_selectTable}
\alias{access_selectTable}
\title{Importar una tabla de Access}
\usage{
access_selectTable(con, tableName)
}
\arguments{
\item{con}{La conexión. Objeto de tipo "RODBC".}

\item{tableName}{El nombre de la tabla.}
}
\value{
Una matriz de datos de tipo tibble.
}
\description{
Importar una tabla de una base de datos Access ya conectada.
}
\examples{
library(RODBC)
# ruta de la base de datos
db <- "C:/Users/34673/Downloads/Liquidaciones2021.mdb"
# generamos la conexión
con <- access_import(db)

# vemos qué tablas hay:
access_searchTables(con)

# importamos varias bases de datos:
tb_CCAA <- access_selectTable(con, 'tb_CCAA')
tb_CCAA

tb_Provincias <- access_selectTable(con, 'tb_Provincias')
tb_Provincias

}
