% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_2dummy.R
\name{association_2dummy}
\alias{association_2dummy}
\title{Tidy test of proportion}
\usage{
association_2dummy(x, y, ..., name_x = NULL, name_y = NULL)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{...}{Optional. Parameters of \code{prop.test} function.}

\item{name_x}{Optional. Name of the x variable (character).}

\item{name_y}{Optional. Name of the y variable (character).}
}
\value{
A data.frame with the results.
}
\description{
Carry out a proportion test for two dummy variables and export it in a tidy
format with \code{broom}. It is also calcul the Cramer V.
}
\examples{
association_2dummy(mtcars$am,mtcars$vs)

}
