% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_2fac.R
\name{association_2fac}
\alias{association_2fac}
\title{Tidy test of chi-squared contingency table test}
\usage{
association_2fac(x, y, ..., name_x = NULL, name_y = NULL)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{...}{Optional. Parameters of \code{chisq.test} function.}

\item{name_x}{Optional. Name of the x variable (character).}

\item{name_y}{Optional. Name of the y variable (character).}
}
\value{
A data.frame with the results.
}
\description{
Carry out a chi-squared contingency table test for two vectors and export it in a tidy
format with \code{broom}.
}
\examples{
association_2fac( as.factor(mtcars$cyl), as.factor(mtcars$am) )
}
