% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_2num_difmean.R
\name{association_2num_difmean}
\alias{association_2num_difmean}
\title{Tidy test of mean difference}
\usage{
association_2num_difmean(x, y, ..., name_x = NULL, name_y = NULL)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{...}{Optional. Parameters of \code{t.test} function.}

\item{name_x}{Optional. Name of the x variable (character).}

\item{name_y}{Optional. Name of the y variable (character).}
}
\value{
A data.frame with the results.
}
\description{
Carry out a diference mean test for two vectors and export it in a tidy
format with \code{broom}.
}
\examples{
association_2num_cor(mtcars$mpg, mtcars$disp)
}
