% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_multiple_dif_test.R
\name{association_multiple_dif_test}
\alias{association_multiple_dif_test}
\title{Estimate the differences in means between each level of a factor}
\usage{
association_multiple_dif_test(df, pairs_to_check, pesos = NULL, ...)
}
\arguments{
\item{df}{A data.frame}

\item{pairs_to_check}{A named list. Each element a character vector with factor names. Each element name is the name of the numeric variable.}

\item{pesos}{Optional. A vector of weights.}

\item{...}{Other arguments passed for \code{modelbased::estimate_contrasts}}
}
\value{
A list.
}
\description{
Carry out a multiple comparison test. By default Holm method is used.
}
\examples{
df <- mtcars
df$cyl <- as.factor(df$cyl)
df$gear <- as.factor(df$gear)
comparison <- list('mpg'  = c('cyl','gear'),
                   'disp' = c('cyl','gear') )

association_multiple_dif_test( df,pairs_to_check = gear )
}
