% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_num_dummy.R
\name{association_num_dummy}
\alias{association_num_dummy}
\title{Tidy t-test}
\usage{
association_num_dummy(x, y, ..., name_x = NULL, name_y = NULL)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{...}{Optional. Parameters of \code{t.test} function.}

\item{name_x}{Optional. Name of the x variable (character).}

\item{name_y}{Optional. Name of the y variable (character).}
}
\value{
A data.frame with the results.
}
\description{
Carry out a t-test for two vectors and export it in a tidy
format with \code{broom}.
}
\examples{
association_num_dummy( mtcars$mpg, mtcars$vs  )
association_num_dummy( mtcars$vs, mtcars$mpg  ) # order is not important
}
