% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_num_fac.R
\name{association_num_fac}
\alias{association_num_fac}
\title{Tidy analysis of variance test (ANOVA)}
\usage{
association_num_fac(x, y, ..., name_x = NULL, name_y = NULL)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{...}{Optional. Parameters of \code{aov} function.}

\item{name_x}{Optional. Name of the x variable (character).}

\item{name_y}{Optional. Name of the y variable (character).}
}
\value{
A data.frame with the results.
}
\description{
Carry out a ANOVA test for two vectors and export it in a tidy
format with \code{broom}.
}
\examples{
association_num_fac(mtcars$mpg, as.factor(mtcars$cyl))
}
