% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_italic.R
\name{bold_italic}
\alias{bold_italic}
\title{Bold or italicize an expression (html)}
\usage{
bold_italic(x, which = c("bold", "italic", "both"), ...)
}
\arguments{
\item{x}{A character vector of length 1.}

\item{which}{A character vector of length 1. Options: 'bold', 'italic' or 'both'.}

\item{...}{Other parameters passed onto \code{glue::glue} function.}
}
\value{
An html character vector.
}
\description{
Vector changes from character to html character. Designed to be used in shiny or html outputs.
}
\examples{
bold_italic('Hola')
bold_italic(c('Hola'),'italic')
bold_italic('Hola','both')
}
