% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalan_month_label.r
\name{catalan_month_label}
\alias{catalan_month_label}
\title{Add Month labels in Catalan to ggplot2}
\usage{
catalan_month_label(breaks)
}
\arguments{
\item{breaks}{this argument is specified inside \code{scale_x_date}. See Examples.}
}
\value{
A ScaleContinuousDate object for a ggplot.
}
\description{
Converting English default month labels to calatan ones in ggplots.
}
\examples{
df <- data.frame(
   x = seq.Date(lubridate::as_date('2020-01-01'),lubridate::as_date('2022-05-01'),
                by = '1 month'),
   y = rnorm(29) )

ggplot(df, aes(x,y))+
 geom_line()+
 scale_x_date(breaks = '1 month', # specify step width
              labels = catalan_month_label ) # add the function to label param
}
