% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_ksnet.R
\name{color_ksnet}
\alias{color_ksnet}
\title{Apply an ideas42 color scheme to your plots and figures}
\usage{
color_ksnet(name, n, type = c("discrete", "continuous"))
}
\value{

}
\description{
Apply an ideas42 color scheme to your plots and figures
}
\examples{

}
