% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_and_share.R
\name{count_and_share}
\alias{count_and_share}
\title{Count and Share}
\usage{
count_and_share(df, ...)
}
\arguments{
\item{df}{Un tibble o dataframe.}

\item{...}{Nombre de las variables a utilizar.}
}
\value{
Un nuevo tibble con el conteo y el porcentaje
}
\description{
Tras realizar un group_by es habitual calcular el porcentaje.
Con esta función se obtiene por defecto.
}
\examples{

mtcars \%>\%
 count_and_share(cyl, vs)
}
