% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimalplaces.r
\name{decimalplaces}
\alias{decimalplaces}
\title{Number of decimas: internal
Internal function for generate_synthetic_object. Not exported.}
\usage{
decimalplaces(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Number of decimas: internal
Internal function for generate_synthetic_object. Not exported.
}
