% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_pair_type.R
\name{detect_pair_type}
\alias{detect_pair_type}
\title{Type of pair of variables}
\usage{
detect_pair_type(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
A character vector of length 1. Values can be '2dummies','2num',
'2fac','num_dummy' or 'num_fac'.
}
\description{
Detect the type of a pair of variabes.
}
\examples{

detect_pair_type(as.factor(mtcars$carb),as.factor(mtcars$cyl)) # '2fac'
detect_pair_type(mtcars$am,  mtcars$vs) # '2dummies
detect_pair_type(mtcars$mpg, mtcars$disp) # '2num'
detect_pair_type(mtcars$mpg, mtcars$vs) # 'num_dummy'
detect_pair_type(mtcars$mpg, as.factor(mtcars$cyl)) # 'num_fac'
}
