% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_association.r
\name{extract_association}
\alias{extract_association}
\title{Extract specifics tests from \code{ksnet::get_association}}
\usage{
extract_association(obj, which = c("ANOVA", "corr", "t_test", "chisq_test"))
}
\arguments{
\item{obj}{a dataframe or a list.}

\item{which}{One of 'ANOVA','corr','t_test','chisq_test'.}
}
\value{
A data.frame
}
\description{
This function is a helper to work with \code{ksnet::get_association}. After obtaining
the results, for example, we could be interesting in keep in one data.frame only those
related to either ANOVA or ANOVA and chi 2.
}
\examples{

mtcars$cyl <- factor(mtcars$cyl)

mtcars_asociacion <- ksnet::get_get_association(mtcars)

extract_association( mtcars_asociacion, 'ANOVA' )



}
