% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_catastro.r
\name{extract_from_catastro}
\alias{extract_from_catastro}
\title{Extract data from catastro object}
\usage{
extract_from_catastro(list_scrape, which = c("parcelas", "inmuebles"))
}
\arguments{
\item{list_scrape}{A catastro object (the output of \code{scrape_catastro}).}

\item{which}{Options: 'parcelas', 'inmuebles' or c('parcelas','inmuebles').}
}
\value{
A tibble with the specified information.
}
\description{
After scraping data with \code{scrape_catastro} is obtained an object of class catastro which
is a list with at least two elements (\code{parcelas} and \code{inmuebles}). With this function
that information can be translated into a tibble: both the \code{parcelas} and \code{inmuebles}
as well as both
}
\examples{

# Real example

urls <- c("https://www1.sedecatastro.gob.es/CYCBienInmueble/OVCListaBienes.aspx?rc1=9797905&rc2=VK3799F",
          "https://www1.sedecatastro.gob.es/CYCBienInmueble/OVCListaBienes.aspx?rc1=2831834&rc2=VK4723B")

address <- c('CL ARMENTEROS 44','CL ABTAO 30')

# web scraping:
example_catastro <- scrape_catastro(urls,address)

# extract information:

extract_from_catastro(example_catastro, which = 'inmuebles')

extract_from_catastro(example_catastro, which = 'parcelas')

extract_from_catastro(example_catastro, c('inmuebles','parcelas'))

}
