% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_case_when.R
\name{fct_case_when}
\alias{fct_case_when}
\title{Obtain a factor after a \code{case_when}}
\usage{
fct_case_when(...)
}
\arguments{
\item{...}{Input of a \code{case_when}.}
}
\value{
A factor vector.
}
\description{
The output of the \code{case_when} function is a character, which means that when
you want to work with factors you have to apply the \code{factor} function afterwards.
}
\details{
With this function the usual \code{case_when} procedure is followed but a factor is
obtained with the levels ordered according to the appearance in
the \code{case_when} input.
}
\examples{

# Comparing the outputs of the two functions:

library(dplyr)

as_tibble(mtcars) \%>\%
    select( cyl ) \%>\%
    mutate(
     cyl_case_when = case_when(
         cyl == 4 ~ 'Bajo',
         cyl == 6 ~ 'Medio',
         cyl == 8 ~ 'Alto' ),

     cyl_fct_case_when = fct_case_whe
         cyl == 4 ~ 'Bajo'
         cyl == 6 ~ 'Medio'
         cyl == 8 ~ 'Alto' )
     )

}
