% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_object.r
\name{generate_synthetic_object}
\alias{generate_synthetic_object}
\alias{generate_synthetic_object.default}
\alias{generate_synthetic_object.numeric}
\alias{generate_synthetic_object.data.frame}
\title{Generate synthetic vectors from their sampling distributions}
\usage{
generate_synthetic_object(obj, seed = NULL, n_news = NULL)

\method{generate_synthetic_object}{default}(obj, seed = NULL, n_news = NULL)

\method{generate_synthetic_object}{numeric}(obj, seed = NULL, n_news = NULL)

\method{generate_synthetic_object}{data.frame}(obj, seed = NULL, n_news = NULL)
}
\arguments{
\item{obj}{A dataframe, numeric vector or character/factor vector.}

\item{seed}{Specify seed when replication is desired.}

\item{n_news}{Length of the new vectors. By default, same length as input.}
}
\value{
An object of equal dimensions (if \code{n_news} is not specified) with synthetic values.
}
\description{
One way of creating synthetic data is replicating sampling distribution
of the variable. This function can be applied to dataframes, numeric or
character/factor vectors and obtain an object of equal length with synthetic values.
}
\examples{
generate_synthetic_object(mtcars)

generate_synthetic_object(mtcars$mpg)

generate_synthetic_object(as.factor(mtcars$cyl))
}
