% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSave.R
\name{ggSave}
\alias{ggSave}
\title{Exportar los gráficos}
\usage{
ggSave(
  p,
  name,
  ext = ".png",
  width = 8000,
  height = 4500,
  unit = "px",
  dpi = 720,
  type = "cairo",
  ...
)
}
\arguments{
\item{p}{Un gráfico ggplo2.}

\item{name}{Character. La ruta y nombre del gráfico.}

\item{ext}{Character. La extensión. Por defecto '.png'.}

\item{width}{Numeric. Por defecto 8000.}

\item{height}{Numeric. Por defecto 4500.}

\item{unit}{Character. Por defecto "px".}

\item{dpi}{Numeric. Por defecti 720.}

\item{type}{Character. Por defecto "cairo".}

\item{...}{Otros argumentos que reciba la función \code{ggplot2::ggsave}.}
}
\value{
Exporta el gráfico.
}
\description{
Define los parámetros de \code{ggplot2::ggsave}.
}
\examples{

p <- ggplot2::ggplot( data = mtcars, ggplot2::aes( disp, mpg ) ) + ggplot2::geom_point()

ggSave( p, 'Graficos/migrafico' )

}
