% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stata_format.r
\name{guessAndset_class}
\alias{guessAndset_class}
\title{Adivina y modifica la clase de un vector}
\usage{
guessAndset_class(x, setlevels = NULL)
}
\arguments{
\item{x}{Un vector.}

\item{setlevels}{Los niveles de un posible factor. Por defecto NULL.}
}
\value{
Un vector con la clase 'adivinada'.
}
\description{
Internamente recurre a \code{vroom::guess_type}.
}
