% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_by_sampling_distribution.r
\name{impute_by_sampling_distribution}
\alias{impute_by_sampling_distribution}
\title{Imput missing values by variable samplingd distribution}
\usage{
impute_by_sampling_distribution(obj, set_seed = NULL)
}
\arguments{
\item{obj}{A dataframe, numeric vector or character/factor vector.}

\item{seed}{Specify seed when replication is desired.}
}
\value{
The same object without NA values as they have been imputed.
}
\description{
This function take advantage of \code{generate_synthetic_object} to impute
missing data. Read \code{help(generate_synthetic_object)} for more information.
}
\examples{

impute_random(c(mtcars$mpg,NA,NA,NA,NA,NA), seed = 123)

data_temp <- data.frame(
    x = c(mtcars$mpg,NA,NA,NA,NA,NA),
    y = c(mtcars$cyl,NA,NA,NA,NA,NA))

as.data.frame(impute_random(data_temp,set_seed = 123) )

dplyr::mutate(data_temp, x_impute = impute_random(x))
}
