% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_2dummies.R
\name{is_2dummies}
\alias{is_2dummies}
\title{Pair of dummy variables}
\usage{
is_2dummies(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
A logical value.
}
\description{
Detect if two vetors are both dummies.
}
\examples{
is_2dummies(mtcars$vs, mtcars$am)
}
