% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_2fac.R
\name{is_2fac}
\alias{is_2fac}
\title{Pair of factor variables}
\usage{
is_2fac(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
A logical value.
}
\description{
Detect if two vetors are both factors (or characters).
}
\examples{
is_2fac(as.factor(mtcars$cyl), as.factor(mtcars$gear))

is_2fac(as.character(mtcars$cyl), as.character(mtcars$gear))
}
