% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_2num.R
\name{is_2num}
\alias{is_2num}
\title{Pair of numeric variables}
\usage{
is_2num(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
A logical value.
}
\description{
Detect if two vetors are both numerics.
}
\examples{
is_2num(mtcars$mpg, mtcars$disp)
}
