% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dummy.R
\name{is_dummy}
\alias{is_dummy}
\title{Dummy vectors}
\usage{
is_dummy(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A logical value.
}
\description{
Detect if a vetor only has 2 unique values, regardless of being numeric, factor or logical.
}
\examples{

is_dummy(mtcars$vs)
}
