% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_num_dummy.R
\name{is_num_dummy}
\alias{is_num_dummy}
\title{Pair of one numeric variable and one dummy}
\usage{
is_num_dummy(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
A logical value.
}
\description{
Detect if one vector is numeric and the another one a dummy.
}
\examples{
is_num_dummy(mtcars$mpg, mtcars$vs)

}
