% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_num_fac.R
\name{is_num_fac}
\alias{is_num_fac}
\title{Pair of one numeric variable and one factor}
\usage{
is_num_fac(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
A logical value.
}
\description{
Detect if one vector is numeric and the another one a factor (or character).
}
\examples{
is_num_dummy(mtcars$mpg, as.factor(mtcars$vs))
}
