% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksnet_bar.R
\name{ksnet_bar}
\alias{ksnet_bar}
\title{Generate a bar chart}
\usage{
ksnet_bar(data, xvar, yvar, fill = color_ksnet("ksnet_classic")[1])
}
\arguments{
\item{data}{data.frame}

\item{xvar}{variable}

\item{yvar}{variable}
}
\value{

}
\description{
This function generates a ggplot2 object, so it's easily
expanded with labels, themes, and other elements.
}
\examples{

x <- c("A","B","C","D")
y <- c(5, 8, 12,15)
my_data <- data.frame(x = x, y = y)
ksnet_bar(my_data, x, y)
}
