% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksnet_hist.R
\name{ksnet_hist}
\alias{ksnet_hist}
\title{Generate a histogram with ksnet styling}
\usage{
ksnet_hist(data, xvar, fill = color_ksnet()[1])
}
\arguments{
\item{data}{data.frame}

\item{xvar}{variable}
}
\value{
ggplot2
}
\description{
Explore the distribution of your data with a histogram. This function generates
a ggplot2 object, so it's easily expanded.
}
\examples{

my_data <- data.frame(x = c(1,2,3,4,5))
ksnet_hist(my_data, x)
}
