% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksnet_table.R
\name{ksnet_table}
\alias{ksnet_table}
\title{Tabla al estilo de KSNET (con gt)}
\usage{
ksnet_table(df)
}
\arguments{
\item{df}{dataframe para hacer la tabla}
}
\description{
Tabla al estilo de KSNET (con gt)
}
\examples{
x <- c("Barcelona", "Madrid", "Murcia")
y <-  c(115, 90, 145)
z  <- c(2019, 2020, 2019)
mis_datos <- data.frame(provincia = x, variable = y, fecha = z)
ksnet_table(mis_datos)

}
