% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_search_interactions.r
\name{lm_search_interactions}
\alias{lm_search_interactions}
\title{Search for interactions in linear models}
\usage{
lm_search_interactions(
  data,
  vd,
  vars_interactions,
  controls = NULL,
  n_muestras = 10,
  p_muestras = 0.9,
  var_to_stratify = NULL,
  peso = NULL,
  get_summary = TRUE
)
}
\arguments{
\item{data}{A dataframe.}

\item{vd}{Name of the dependent variable (character).}

\item{vars_interactions}{A character vector with the names of variables to interact betwean pairs.}

\item{controls}{A character vector with the names of variable controls (if desired) or 'all'. Default is NULL. See details for more information.}

\item{n_muestras}{Number of samples to take from the data. If 1 is speficied, it takes the the entire dataset.}

\item{p_muestras}{Proportion of observations to keep in each sample. Only if n_muestras greater than 1.}

\item{var_to_stratify}{Name of the variable to stratify when creating samples. Default is NULL.}

\item{peso}{Numeric vector with sampling weights. If NULL is used a vector of ones.}

\item{get_summary}{Logical. It should be returned the coefficient of every sample or just the average? Default if TRUE.}
}
\value{
A list with two elements: 'outcome' is a dataframe with the results, and 'formulas' showing the formulas used.
}
\description{
Method imported from Machine Learning techniques but with inferential statistics in mind.
Detect wich pair of interaction if statistically significant 'regardless' input data.
}
\details{
When specifying 'all' inf 'controls', all variables present in vars_interactions will be used in every linear model.
}
\examples{
res_temp <- lm_search_interactions( data = mtcars,
                                    vd = 'mpg',
                                    vars_interactions = c('cyl','disp','vs'),
                                    controls = 'all',
                                    n_muestras = 3,
                                    p_muestras = .9,
                                    get_summary = F)
res_temp$outcome
res_temp$formulas
}
