% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_samples.r
\name{make_samples}
\alias{make_samples}
\title{Generate samples from a dataset}
\usage{
make_samples(
  data,
  n_samples,
  p,
  var_to_stratify = NULL,
  pond = NULL,
  seed = 12
)
}
\arguments{
\item{data}{a dataframe}

\item{n_samples}{Number of samples.}

\item{p}{Numeric between 0 and 1. Percentage of observations to keep in each sample.}

\item{var_to_stratify}{If desired, name of variable to straty the sampling. Default NULL.}

\item{pond}{A numeric vector of sample weights. Default NULL.}

\item{seed}{Numeric.}
}
\value{
A list with dataframes.
}
\description{
Generate samples of a dataset.
}
\examples{

ecv <- read_csv('ecv.csv')

# Make 10 samples, 90\% of observations in each one and stratify by CCAA
ecv_samples <- make_samples(ecv, 10, .9, 'CCAA')

# Taking into account sample weights
ecv_samples <- make_samples(ecv, 10, .9, 'CCAA', ecv$factor_hogar)

}
