% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_ksnet.R
\name{scale_color_ksnet_continuous}
\alias{scale_color_ksnet_continuous}
\title{Paleta continua de colores de KSNET (color)}
\usage{
scale_color_ksnet_continuous(palette = "verde", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Nombre de la paleta ("verde", "gris", "amarillo", "rojo",
"verde_amarillo", "rojo_amarillo")}

\item{reverse}{Booleano para revertir la paleta (TRUE o FALSE)}

\item{...}{Argumentos adicionales}
}
\description{
Paleta continua de colores de KSNET (color)
}
\examples{
library(ggplot2)
library(dplyr)
x <- rnorm(100)
y <-  rnorm(100, x, 1)
z <-  rnorm(100)
mis_datos <- data.frame(variable = x, valor = y, valor2 = z)
mis_datos \%>\%
ggplot(aes(x = variable, y = valor, color = valor2)) +
geom_point() +
scale_color_ksnet_continuous() +
theme_ksnet()
}
