% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_ksnet.R
\name{scale_fill_ksnet_continuous}
\alias{scale_fill_ksnet_continuous}
\title{Paleta continua de colores de KSNET (fill)}
\usage{
scale_fill_ksnet_continuous(palette = "verde", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Nombre de la paleta ("verde", "gris", "amarillo", "rojo",
"verde_amarillo", "rojo_amarillo")}

\item{reverse}{Booleano para revertir la paleta (TRUE o FALSE)}

\item{...}{Argumentos adicionales}
}
\description{
Paleta continua de colores de KSNET (fill)
}
