% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_catastro.R
\name{scrape_catastro}
\alias{scrape_catastro}
\title{Web Scraping of Catastro}
\usage{
scrape_catastro(url, ID = NA_character_, i = NULL, verbose = FALSE)
}
\arguments{
\item{url}{A character vector of variable length with the url.}

\item{ID}{An optional character vector with the unique identification of each real state. For example, its name or address.}

\item{i}{An internal argument when url vector is of length greater than 1. Do not use this parameter.}

\item{verbose}{If TRUE, the iteration number is printed on the console. Default set to FALSE.}
}
\value{
A list with two elements: the first one with the information of the main real state (\code{parcela}).
In the second element is the information regarding real states associated with the main one (\code{inmuebles}).
}
\description{
This function is designed to extract the information related to
real estates from Catastro.
}
\examples{

# Real examples

url_1 <- "https://www1.sedecatastro.gob.es/CYCBienInmueble/OVCListaBienes.aspx?rc1=9797905&rc2=VK3799F"
url_2 <- "https://www1.sedecatastro.gob.es/CYCBienInmueble/OVCListaBienes.aspx?rc1=2831834&rc2=VK4723B"

address_1 <- 'CL ARMENTEROS 44'
address_2 <- 'CL ABTAO 30'

# web scraping the information of the first one:

scrape_catastro( url_1, address_1 )

# web scraping the information of the two webpages:

urls <- c(url_1,url_2)
address <- c(address_1,address_2)

scrape_catastro( urls, address, verbose = TRUE )

}
