% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_fac_fac.R
\name{summarise_fac_fac}
\alias{summarise_fac_fac}
\title{Extract descriptive statistics between two factors}
\usage{
summarise_fac_fac(x, y, peso = NULL, name_x = NULL, name_y = NULL, round_p = 3)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{peso}{NOT IMPLEMENTED.}

\item{name_x}{Optional. Name of the x variable (character).}

\item{name_y}{Optional. Name of the Y variable (character).}

\item{round_p}{A numeric of length 1. Number of decimals of proportions.}
}
\value{
A data.frame
}
\description{
Extract descriptive statistics between two factors
}
\examples{
# DATA FOR THE EXAMPLES #

library(dplyr)
df <- mtcars \%>\%
 filter( carb != 6 & carb != 8 ) \%>\%
 mutate( across(c(cyl,vs,am,gear, carb), as.factor) )

df$vs <- as.factor(df$vs)
summarise_fac_fac( df$cyl, df$vs, peso = abs(runif(30,0.5,3)) )
}
