% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_fac_num.R
\name{summarise_fac_num}
\alias{summarise_fac_num}
\title{Extract descriptive statistics grouping by a factor}
\usage{
summarise_fac_num(
  x,
  y,
  peso = NULL,
  name_x = NULL,
  name_y = NULL,
  round_stats = 3,
  round_p = 3
)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{peso}{Optional. A vector of weights.}

\item{name_x}{Optional. Name of the x variable (character).}

\item{name_y}{Optional. Name of the y variable (character).}

\item{round_stats}{A numeric of length 1. Number of decimals of stats.}

\item{round_p}{A numeric of length 1. Number of decimals of proportions.}
}
\value{
A data.frame.
}
\description{
Extract descriptive statistics grouping by a factor
}
\examples{

# DATA FOR THE EXAMPLES #

library(dplyr)
df <- mtcars \%>\%
 filter( carb != 6 & carb != 8 ) \%>\%
 mutate( across(c(cyl,vs,am,gear, carb), as.factor) )

summarise_fac_num(df$mpg, df$cyl,runif(30,0.5,1.5))
summarise_fac_num(mtcars$mpg, as.factor(mtcars$cyl))
summarise_fac_num(mtcars$mpg, mtcars$am)
}
