% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{summary_statistics}
\alias{summary_statistics}
\title{Get (weighted) summary statistics}
\usage{
summary_statistics(df, var, peso = NULL)
}
\arguments{
\item{df}{A dataframe.}

\item{var}{Variable name.}

\item{peso}{A numeric vector of weights. Default is NULL.}
}
\value{
A dataframe.
}
\description{
This function computes the summary statistics of \code{x}.
}
\examples{

summary_statistics(mtcars, disp)

mtcars \%>\%
 group_by(cyl) \%>\%
 summary_statistics(mpg)
}
