% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_statsFac}
\alias{summary_statsFac}
\title{Get (weighted) summary statistics of factor variables}
\usage{
summary_statsFac(df, var, peso = NULL)
}
\arguments{
\item{df}{A data.frame.}

\item{var}{Character. Variable/s name.}

\item{peso}{Character.A numeric vector of weights. Default is NULL.}
}
\value{
A dataframe.
}
\description{
This function computes the summary statistics of one or more variables.
It can receives a grouped data.frame.
}
\examples{

library(dplyr)

mtcars \%>\%
 group_by( cyl ) \%>\%
 summary_statsFac( c('vs') )

 mtcars \%>\%
     mutate( 'peso' = runif(32) ) \%>\%
     group_by( cyl ) \%>\%
     summary_statsFac( 'vs', peso = 'peso' )

 mtcars \%>\%
     group_by( cyl, am ) \%>\%
     summary_statsFac( c('vs') )

 mtcars \%>\%
     group_by( cyl ) \%>\%
     summary_statsFac( c('am','vs') )
}
