% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_statsNum}
\alias{summary_statsNum}
\title{Get (weighted) summary statistics of numeric variables}
\usage{
summary_statsNum(df, var, peso = NULL)
}
\arguments{
\item{df}{A data.frame.}

\item{var}{Character. Variable/s name.}

\item{peso}{Character.A numeric vector of weights. Default is NULL.}
}
\value{
A dataframe.
}
\description{
This function computes the summary statistics of one or more variables.
It can receives a grouped data.frame.
}
\examples{

library(dplyr)

summary_statsNum(mtcars, 'disp')

mtcars \%>\%
 group_by(cyl) \%>\%
 summary_statsNum('mpg')

 mtcars \%>\%
     mutate( 'peso' = runif(32) ) \%>\%
     group_by( cyl ) \%>\%
     summary_statsNum(c('mpg','disp'), peso = 'peso')
}
