% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_resumen.R
\name{tabla_resumen}
\alias{tabla_resumen}
\title{Convertir un objeto table creado con la función summary() a un data.frame}
\usage{
tabla_resumen(t)
}
\arguments{
\item{t}{table creado con la función summary aplicada sobre un data.frame}
}
\value{
Un df en formato largo con tres columnas: variable (nombre de cada columna para la cual hacemos un estadístico), estadístico y valor.
}
\description{
Esta función convierte un objeto table creado con la función summary() a un data.frame en formato largo
Esto nos permite trabajar con los estadísticos descriptivos de una o varias variables utilizando los métodos para data.frame de R Base o del tidyverse.
}
\examples{
df <- data.frame(a = c(1,2,3,4), b = c(5,6,7,8))
sum <- summary(df)
tabla_resumen(sum)
}
