% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ksnet_light.r
\name{theme_ksnet_light}
\alias{theme_ksnet_light}
\title{Theme KSNET fot ggplots}
\usage{
theme_ksnet_light(
  legend_position = "right",
  title_size = 14.5,
  subtitle_size = 12.5,
  x_size = 12.5,
  y_size = 12.5,
  axis_x_size = 12,
  axis_y_size = 12,
  x_vjust = 4,
  remove_grid_major_y = TRUE,
  remove_grid_major_x = TRUE,
  is_x_date = FALSE,
  plot_font = "Public Sans"
)
}
\arguments{
\item{legend_position}{character vector. Default is right.}

\item{title_size}{numeric. Default 14.5.}

\item{subtitle_size}{numeric. Default 12.5.}

\item{x_size}{numeric. Axis title size. Default 12.5.}

\item{y_size}{numeric. Axis title size. Default 12.5.}

\item{axis_x_size}{numeric. Axis text size. Default 12.}

\item{axis_y_size}{numeric. Axis text size. Default 12.}

\item{x_vjust}{numeric. Default 4.}

\item{remove_grid_major_y}{logical. Default TRUE.}

\item{remove_grid_major_x}{logical. Default TRUE.}

\item{is_x_date}{when x axis represents a date vector set to TRUE to adjust vjust.}

\item{font_family}{default is Public Sants. But is is necessary to have it installed. See Details.}
}
\value{
An object of class gg theme to add for an existing ggplot object.
}
\description{
This is one of the available ksnet themes for plots. Its parameters
are just those specified within \code{theme} function. When an argument
is not accesible with this function, it can be combined with another
specific \code{theme}.
}
\details{
At the beginning of your script, after having installed Public Sans font, you should
write the following code to load the font:
windowsFonts('Public Sans'='Public Sans')
}
\examples{
library(ggplot2)

ggplot(mtcars,aes(mpg))+
  geom_density()+
  theme_ksnet_light()

ggplot(mtcars,aes(mpg))+
  geom_density()+
  theme_ksnet_light(remove_grid_major_x = FALSE)


}
