% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_ksnet.R
\name{color_ksnet}
\alias{color_ksnet}
\title{Aplicar un color KSNET a las figuras}
\usage{
color_ksnet(
  palette = "classic",
  reverse = FALSE,
  n,
  type = c("discrete", "continuous"),
  ...
)
}
\value{

}
\description{
Aplicar un color KSNET a las figuras
}
\examples{

}
