% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksnet_scatter.R
\name{ksnet_scatter}
\alias{ksnet_scatter}
\title{Generate a scatterplot}
\usage{
ksnet_scatter(data, xvar, yvar, color = color_ksnet()[1])
}
\arguments{
\item{data}{data.frame}

\item{xvar}{variable}

\item{yvar}{variable}
}
\value{
ggplot
}
\description{
Explore the distribution of your data with a scatterplot This function generates
a ggplot2 object, so it's easily expanded with labels, themes, and other elements.
}
\examples{

x <- rnorm(100)
y <-  rnorm(100, x, 1)
my_data <- data.frame(x = x, y = y)
ksnet_scatter(my_data, x, y)
}
