% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksnet_table.R
\name{ksnet_table}
\alias{ksnet_table}
\title{Tabla al estilo de KSNET (con gt)}
\usage{
ksnet_table(df)
}
\arguments{
\item{df}{dataframe para hacer la tabla}
}
\description{
Tabla al estilo de KSNET (con gt)
}
