% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_ksnet.R
\name{scale_color_ksnet_discrete}
\alias{scale_color_ksnet_discrete}
\title{Paleta discreta de colores de KSNET (color)}
\usage{
scale_color_ksnet_discrete(palette = "classic", reverse = FALSE, ...)
}
\arguments{
\item{palette}{El nombre de la paleta ("classic", "bright")}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Argumentos adicionales}
}
\description{
Paleta discreta de colores de KSNET (color)
}
\examples{
x <- c(1,2,3,4,5)
y <-  rnorm(5, x, 0.5)
mis_datos <- data.frame(x = x, y = y)
mis_datos \%>\%
ggplot(aes(x = x, y = y, color = y)) +
geom_point()
scale_color_ksnet_discrete() +
theme_ksnet()

}
