% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_ksnet.R
\name{scale_color_ksnet_discrete}
\alias{scale_color_ksnet_discrete}
\title{Paleta discreta de colores de KSNET (color)}
\usage{
scale_color_ksnet_discrete(palette = "classic", reverse = FALSE, ...)
}
\arguments{
\item{palette}{El nombre de la paleta ("classic", "bright")}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Argumentos adicionales}
}
\description{
Paleta discreta de colores de KSNET (color)
}
\examples{
library(ggplot2)
library(dplyr)
x <- c("A","B","C","D")
y <- c(5, 8, 12,15)
mis_datos <- data.frame(variable = x, valor = y)
mis_datos \%>\%
ggplot(aes(x = variable, y = valor, color = variable)) +
geom_point() +
scale_color_ksnet_discrete() +
theme_ksnet()

}
