% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_ksnet.R
\name{scale_fill_ksnet_discrete}
\alias{scale_fill_ksnet_discrete}
\title{Paleta discreta de colores de KSNET (fill)}
\usage{
scale_fill_ksnet_discrete(palette = "classic", reverse = FALSE, ...)
}
\arguments{
\item{palette}{El nombre de la paleta ("classic", "bright")}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Argumentos adicionales}
}
\description{
Paleta discreta de colores de KSNET (fill)
}
